+++ quantify_scheduler/instrument_coordinator/components/qblox.py
@@ -8,7 +8,7 @@ import logging
 from abc import abstractmethod
 from collections import namedtuple
 from dataclasses import dataclass
-from typing import Any, Callable, Dict, Optional, Tuple, Type, Union
+from typing import Any, Callable, Dict, Optional, Tuple, Type, Union, List
 
 import numpy as np
 from qblox_instruments import (
@@ -596,9 +596,9 @@ class QRMComponent(QbloxInstrumentCoordinatorComponentBase):
                 "integration_length_acq",
                 settings.integration_length_acq,
             )
-            self._acquisition_manager.integration_length_acq = (
-                settings.integration_length_acq
-            )
+            self._acquisition_manager.acq_duration[
+                seq_idx
+            ] = settings.integration_length_acq
 
         self._set_parameter(
             self.instrument[f"sequencer{seq_idx}"], "demod_en_acq", settings.nco_en
@@ -761,7 +761,7 @@ class _QRMAcquisitionManager:
         self.acquisition_metadata: AcquisitionMetadata = acquisition_metadata
 
         self.scope_mode_sequencer: Optional[str] = None
-        self.integration_length_acq: Optional[int] = None
+        self.acq_duration: List[Optional[int]] = [None] * number_of_sequencers
         self.seq_name_to_idx_map = {
             f"seq{idx}": idx for idx in range(number_of_sequencers)
         }
@@ -800,14 +800,25 @@ class _QRMAcquisitionManager:
             acquisition_function: Callable = protocol_to_function_mapping[
                 acq_metadata.acq_protocol
             ]
-
+            if (
+                self.acq_duration[seq_idx] < 0
+                or self.acq_duration[seq_idx] > constants.MAX_SAMPLE_SIZE_ACQUISITIONS
+            ):
+                raise ValueError(
+                    "Attempting to retrieve sample of size "
+                    f"{self.acq_duration[seq_idx]} "
+                    f"(maximum allowed sample size: "
+                    f"{constants.MAX_SAMPLE_SIZE_ACQUISITIONS})"
+                )
             # retrieve the raw data from the qrm sequencer module
             acquisitions = self.instrument.get_acquisitions(seq_idx)
             for acq_channel, acq_indices in acq_metadata.acq_indices.items():
                 # the acquisition_function retrieves the right part of the acquisitions
                 # data structure returned by the qrm
                 i_vals, q_vals = acquisition_function(
-                    acquisitions=acquisitions, acq_channel=acq_channel
+                    acquisitions=acquisitions,
+                    acq_channel=acq_channel,
+                    acq_duration=self.acq_duration[seq_idx],
                 )
 
                 # the Qblox compilation backend verifies that the
@@ -883,13 +894,17 @@ class _QRMAcquisitionManager:
         return ch_and_idx
 
     def _get_scope_data(
-        self, acquisitions: dict, acq_channel: int = 0
+        self, acquisitions: dict, acq_duration: int, acq_channel: int = 0
     ) -> Tuple[np.ndarray, np.ndarray]:
         """
         Retrieves the scope mode acquisition associated with an `acq_channel`.
 
         Parameters
         ----------
+        acquisitions
+            The acquisitions dict as returned by the sequencer.
+        acq_duration
+            The duration of the acquisition, used to truncate the data.
         acq_channel
             The acq_channel to get the scope mode acquisition for.
 
@@ -909,12 +924,16 @@ class _QRMAcquisitionManager:
                     f"acq_channel={acq_channel}  was out-of-range."
                 )
         # NB hardware already divides by avg_count for scope mode
-        scope_data_i = np.array(scope_data["path0"]["data"])
-        scope_data_q = np.array(scope_data["path1"]["data"])
+        scope_data_i = np.array(scope_data["path0"]["data"][:acq_duration])
+        scope_data_q = np.array(scope_data["path1"]["data"][:acq_duration])
         return scope_data_i, scope_data_q
 
+    # pylint: disable=unused-argument
     def _get_integration_data(
-        self, acquisitions: dict, acq_channel: int = 0
+        self,
+        acquisitions: dict,
+        acq_duration: int = constants.MAX_SAMPLE_SIZE_ACQUISITIONS,
+        acq_channel: int = 0,
     ) -> Tuple[np.ndarray, np.ndarray]:
         """
         Retrieves the integrated acquisition data associated with an `acq_channel`.
@@ -923,6 +942,8 @@ class _QRMAcquisitionManager:
         ----------
         acquisitions
             The acquisitions dict as returned by the sequencer.
+        acq_duration
+            Not used in this function.
         acq_channel
             The `acq_channel` from which to get the data.
 
@@ -944,7 +965,7 @@ class _QRMAcquisitionManager:
         return i_data, q_data
 
     def _get_integration_amplitude_data(
-        self, acquisitions: dict, acq_channel: int = 0
+        self, acquisitions: dict, acq_duration: int, acq_channel: int = 0
     ) -> Tuple[np.ndarray, np.ndarray]:
         """
         Gets the integration data but normalized to the integration time (number of
@@ -956,6 +977,8 @@ class _QRMAcquisitionManager:
         ----------
         acquisitions
             The acquisitions dict as returned by the sequencer.
+        acq_duration
+            The duration of the acquisition, used to truncate the data.
         acq_channel
             The `acq_channel` from which to get the data.
 
@@ -966,7 +989,7 @@ class _QRMAcquisitionManager:
         data_q
             Array containing Q-quadrature data.
         """
-        if self.integration_length_acq is None:
+        if acq_duration is None:
             raise RuntimeError(
                 "Retrieving data failed. Expected the integration length to be defined,"
                 " but it is `None`."
@@ -975,8 +998,8 @@ class _QRMAcquisitionManager:
             acquisitions=acquisitions, acq_channel=acq_channel
         )
         compensated_data_i, compensated_data_q = (
-            compensated_data_i / self.integration_length_acq,
-            compensated_data_q / self.integration_length_acq,
+            compensated_data_i / acq_duration,
+            compensated_data_q / acq_duration,
         )
         return compensated_data_i, compensated_data_q