+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Merged branches and closed issues
 
 - NV centers - First operation `SpectroscopyOperation` with compilation for Qblox hardware (!471)
+- NV centers - `Reset` operation with compilation for Qblox hardware (!485)
+- QBlox backend, Operations - The phase of the measurement clock is set to zero at the start of each measurement by default (!434, #296)
+- Operations - ShiftClockPhase now uses `phase_shift` as keyword instead of `phase` (!434)
 
 ## 0.9.0 (2022-10-06)